Fixed error in course component
